backend/requirements.txt
alembic==1.15.1
bidict==0.23.1
blinker==1.9.0
click==8.1.8
coloredlogs==15.0.1
Flask==3.1.0
flask-cors==5.0.1
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
greenlet==3.1.1
h11==0.14.0
humanfriendly==10.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
mpmath==1.3.0
numpy==2.2.4
onnxruntime==1.21.0
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
protobuf==6.30.1
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0
SQLAlchemy==2.0.39
sympy==1.13.3
typing_extensions==4.12.2
Werkzeug==3.1.3
wsproto==1.2.0
